Add files via upload
It is a game which has been created by using the python library named Tkinter. In this game user can play the game as many times they want. Two players are needed to play this game. Players can quit the game in between if they don't want to lose otherwise the winner will get the message of their victory or the message of draw on the game screen